Create User entity with roles enum and DTO
Created the User entity to represent a user in the system, including fields such as name, lastname, email, password, registration date, and role. Implemented validations for required fields and password strength. The UserRole enum was added to define user roles (USER, ADMIN, SYSTEM) for role-based access control. Introduced a UserDTO class for data transfer purposes, which includes user information and annotations for API documentation. The User entity is designed to store data in MongoDB, using annotations for field mappings and indexes, with a focus on data validation and security.